back/requirements.txt
asgiref
channels
channels-redis
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
daphne
matplotlib